+++ core/src/test/java/org/seedstack/business/internal/DefaultRepositoryCollectorTest.java
@@ -7,21 +7,28 @@
  */
 package org.seedstack.business.internal;
 
+import com.google.common.collect.Lists;
 import com.google.inject.Key;
 import com.google.inject.TypeLiteral;
 import com.google.inject.name.Names;
 import com.google.inject.util.Types;
 import org.apache.commons.configuration.Configuration;
-import org.assertj.core.api.Assertions;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.internal.util.reflection.Whitebox;
+import org.seedstack.business.domain.AggregateRoot;
+import org.seedstack.business.domain.BaseAggregateRoot;
+import org.seedstack.business.domain.BaseRepository;
 import org.seedstack.business.domain.Repository;
+import org.seedstack.business.internal.strategy.api.BindingStrategy;
 import org.seedstack.business.repositories.fixtures.MyQualifier;
 import org.seedstack.seed.Application;
 
 import java.lang.reflect.Type;
-import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Map;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -32,16 +39,22 @@ public class DefaultRepositoryCollectorTest {
 
     private DefaultRepositoryCollector underTest;
     private Application application;
-
     private TypeLiteral<?> genericInterface = TypeLiteral.get(Types.newParameterizedType(Repository.class, new Type[]{MyAgg.class}));
 
     @Before
     public void before() {
         application = mock(Application.class);
-        underTest = new DefaultRepositoryCollector(new ArrayList<Class<?>>(), new ArrayList<Class<?>>(), application);
+        underTest = new DefaultRepositoryCollector(
+                Lists.<Class<?>>newArrayList(MySubAgg1.class, MySubAgg2.class),
+                Lists.<Class<?>>newArrayList(MyDefaultRepo.class),
+                application
+        );
     }
 
-    private static class MyAgg {
+    @Test
+    public void testCollectSuperclasses() throws Exception {
+        Collection<BindingStrategy> bindingStrategies = underTest.collect();
+        assertThat(((Map<?, ?>) Whitebox.getInternalState(bindingStrategies.iterator().next(), "constructorParamsMap")).size()).isEqualTo(3);
     }
 
     @Test
@@ -52,7 +65,7 @@ public class DefaultRepositoryCollectorTest {
         when(application.getConfiguration(MyAgg.class)).thenReturn(configuration);
         Key<?> key = underTest.defaultRepositoryQualifier(MyAgg.class, genericInterface);
 
-        Assertions.assertThat(key.getAnnotation()).isEqualTo(Names.named("my-qualifier"));
+        assertThat(key.getAnnotation()).isEqualTo(Names.named("my-qualifier"));
     }
 
     @Test
@@ -63,6 +76,61 @@ public class DefaultRepositoryCollectorTest {
         when(application.getConfiguration(MyAgg.class)).thenReturn(configuration);
         Key<?> key = underTest.defaultRepositoryQualifier(MyAgg.class, genericInterface);
 
-        Assertions.assertThat(key.getAnnotationType()).isEqualTo(MyQualifier.class);
+        assertThat(key.getAnnotationType()).isEqualTo(MyQualifier.class);
+    }
+
+    private static class MyAgg extends BaseAggregateRoot<Long> {
+        @Override
+        public Long getEntityId() {
+            return null;
+        }
+    }
+
+    private static class MySubAgg1 extends MyAgg {
+    }
+
+    private static class MySubAgg2 extends MyAgg {
+    }
+
+    public class MyDefaultRepo<A extends AggregateRoot<K>, K> extends BaseRepository<A, K> {
+        @Override
+        public A load(K id) {
+            return null;
+        }
+
+        @Override
+        public boolean exists(K id) {
+            return false;
+        }
+
+        @Override
+        public long count() {
+            return 0;
+        }
+
+        @Override
+        public void clear() {
+
+        }
+
+        @Override
+        public void delete(K id) {
+
+        }
+
+        @Override
+        public void delete(A aggregate) {
+
+        }
+
+        @Override
+        public void persist(A aggregate) {
+
+        }
+
+        @Override
+        public A save(A aggregate) {
+            return null;
+        }
     }
 }